+++ lib/chewy/query/criteria.rb
@@ -1,21 +1,17 @@
 module Chewy
   class Query
     class Criteria
-      STORAGES = [:options, :query, :facets, :filters, :sort, :fields, :types]
+      STORAGES = [:options, :queries, :facets, :filters, :sort, :fields, :types]
 
-      def initialize
-
+      def initialize options = {}
+        @options = options
       end
 
-      def ==(other)
-        storages == other.storages
+      def == other
+        other.is_a?(self.class) && storages == other.storages
       end
 
-      def storages
-        STORAGES.map { |storage| send(storage) }
-      end
-
-      [:options, :query, :facets].each do |storage|
+      [:options, :facets].each do |storage|
         class_eval <<-METHODS, __FILE__, __LINE__ + 1
           def #{storage}
             @#{storage} ||= {}
@@ -23,7 +19,7 @@ module Chewy
         METHODS
       end
 
-      (STORAGES - [:options, :query, :facets]).each do |storage|
+      (STORAGES - [:options, :facets]).each do |storage|
         class_eval <<-METHODS, __FILE__, __LINE__ + 1
           def #{storage}
             @#{storage} ||= []
@@ -47,8 +43,8 @@ module Chewy
         facets.merge!(modifer)
       end
 
-      def update_query(modifer)
-        query.merge!(modifer)
+      def update_queries(modifer)
+        @queries = queries + Array.wrap(modifer).delete_if(&:blank?)
       end
 
       def update_filters(modifer)
@@ -84,7 +80,21 @@ module Chewy
         clone.merge!(other)
       end
 
+      def request_body
+        body = (_request_query || {}).tap do |body|
+          body.merge!(facets: facets) if facets?
+          body.merge!(sort: sort) if sort?
+          body.merge!(fields: fields) if fields?
+        end
+
+        {body: body.merge!(_request_options)}
+      end
+
     protected
+
+      def storages
+        STORAGES.map { |storage| send(storage) }
+      end
 
       def initialize_clone(other)
         STORAGES.each do |storage|
@@ -93,6 +103,79 @@ module Chewy
             value = Marshal.load(Marshal.dump(value))
             instance_variable_set("@#{storage}", value)
           end
+        end
+      end
+
+      def _request_options
+        options.slice(:size, :from, :explain)
+      end
+
+      def _request_query
+        request_filter = _request_filter
+        request_query = _queries_join(queries, options[:query_mode] || :must)
+
+        if request_filter
+          {query: {
+            filtered: {
+              query: request_query ? request_query : {match_all: {}},
+              filter: request_filter
+            }
+          }}
+        elsif request_query
+          {query: request_query}
+        end
+      end
+
+      def _request_filter
+        filter_mode = options[:filter_mode] || :and
+        request_filter = if filter_mode == :and
+          filters
+        else
+          [_filters_join(filters, filter_mode)]
+        end
+
+        _filters_join([_request_types, *request_filter], :and)
+      end
+
+      def _request_types
+        _filters_join(types.map { |type| {type: {value: type}} }, :or)
+      end
+
+      def _queries_join queries, logic
+        queries = queries.compact
+
+        if queries.many?
+          case logic
+          when :dis_max
+            {dis_max: {queries: queries}}
+          when :must, :should
+            {bool: {logic => queries}}
+          else
+            if logic.is_a?(Float)
+              {dis_max: {queries: queries, tie_breaker: logic}}
+            else
+              {bool: {should: queries, minimum_should_match: logic}}
+            end
+          end
+        else
+          queries.first
+        end
+      end
+
+      def _filters_join filters, logic
+        filters = filters.compact
+
+        if filters.many?
+          case logic
+          when :and, :or
+            {logic => filters}
+          when :must, :should
+            {bool: {logic => filters}}
+          else
+            {bool: {should: filters, minimum_should_match: logic}}
+          end
+        else
+          filters.first
         end
       end
     end